start on refactoring clusters of interest editor
